Separate large Rust test suites
